+++ tests/ops/models/test_privacy_notice.py
@@ -1,11 +1,11 @@
 from typing import Generator
 
 import pytest
-from fideslang.models import Cookies as CookieSchema
 from fideslang.validation import FidesValidationError
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
+from fides.api.models.asset import Asset
 from fides.api.models.experience_notices import ExperienceNotices
 from fides.api.models.location_regulation_selections import DeprecatedNoticeRegion
 from fides.api.models.privacy_notice import (
@@ -17,7 +17,7 @@ from fides.api.models.privacy_notice import (
     PrivacyNoticeHistory,
     UserConsentPreference,
 )
-from fides.api.models.sql_models import Cookies, PrivacyDeclaration
+from fides.api.models.sql_models import PrivacyDeclaration
 from fides.api.schemas.language import SupportedLanguage
 
 
@@ -588,7 +588,7 @@ class TestPrivacyNoticeModel:
             (
                 ["marketing.advertising", "third_party_sharing"],
                 [{"name": "test_cookie"}],
-                [CookieSchema(name="test_cookie")],
+                [{"name": "test_cookie"}],
                 "Data uses overlap exactly",
             ),
             (
@@ -600,13 +600,13 @@ class TestPrivacyNoticeModel:
             (
                 ["marketing", "third_party_sharing"],
                 [{"name": "test_cookie"}],
-                [CookieSchema(name="test_cookie")],
+                [{"name": "test_cookie"}],
                 "Privacy notice use more general than system's, so system's data use is under the scope of the notice",
             ),
             (
                 ["marketing.advertising", "third_party_sharing"],
                 [{"name": "test_cookie"}, {"name": "another_cookie"}],
-                [CookieSchema(name="test_cookie"), CookieSchema(name="another_cookie")],
+                [{"name": "test_cookie"}, {"name": "another_cookie"}],
                 "Test multiple cookies",
             ),
             (["marketing.advertising"], [], [], "No cookies returns an empty set"),
@@ -623,7 +623,6 @@ class TestPrivacyNoticeModel:
         system,
     ):
         """Test different combinations of data uses and cookies between the Privacy Notice and the Privacy Declaration"""
-        db.query(Cookies).delete()
         privacy_notice.data_uses = privacy_notice_data_use
         privacy_notice.save(db)
 
@@ -631,19 +630,18 @@ class TestPrivacyNoticeModel:
         assert privacy_declaration.data_use == "marketing.advertising"
 
         for cookie in declaration_cookies:
-            Cookies.create(
-                db,
-                data={
-                    "name": cookie["name"],
-                    "privacy_declaration_id": privacy_declaration.id,
-                    "system_id": system.id,
-                },
-                check_name=False,
+            asset = Asset(
+                name=cookie["name"],
+                asset_type="Cookie",
+                data_uses=["marketing.advertising"],
+                system_id=system.id,
             )
+            db.add(asset)
+        db.commit()
 
-        assert [
-            CookieSchema.model_validate(cookie) for cookie in privacy_notice.cookies
-        ] == expected_cookies, description
+        assert [cookie.name for cookie in privacy_notice.cookies] == [
+            cookie["name"] for cookie in expected_cookies
+        ], description
 
     def test_generate_privacy_notice_key(self, privacy_notice):
         assert (